Ransomware is created with PHP to make technocrats understand how it actually works.
Ransomware is a type of malware that encrypts a victim’s files. The attacker then demands a ransom from the victim to restore access to the data upon payment. While ransomware has technically been around since the 90’s, its only in the past five years or so that it has really taken off, largely because of the availability of untraceable payment methods like Bitcoins.
Here I’ve created a PHP file that will encrypt all the files and directories from the server’s root directory and only inside the server’s web root directory.
It is purely made for educational purpose to make technocrats understand how the ransomware actually works.
It is tested on the following:
Operating System: Windows 10 Pro (64-bit)
Web Server: XAMPP v7.3.7(64-bit)